refactor(hackerpacks): Extract duplicated code into component
Hackerpack content pages were being duplicated on the index.tsx and [id].tsx
pages. It's been extracted to HackerpackDisplay.tsx and simply included as a
component in the main pages.
